Add string_utils.c and string_utils.h
