+++ src/objects/class.c
@@ -5,6 +5,7 @@
 #include <stdio.h>
 #include <string.h>
 #include "gc.h"
+#include "string_utils.h"
 #include "return_value.h"
 
 #include "objects/str.h"
